ComposeMessage: Improve performance with large groups
Messaging has a number of speed problems when trying
to create messages to large groups. These problems
sometimes cause ANR's if the user tries to use the app
while it is still preparing the message.

The issue is that the actions taken while preparing this
message are very intense in several areas:

1) Listeners are added several times to the
recipient field (init, start, and resume)

2) Lists are turned into huge strings

3) Update functions are used excessively

4) Listeners are added to each contact item several
times.

By removing the redundant listeners
and preventing list to string concatenation we
speed up the process seen after "Adding contacts"
dialog is removed.

Change-Id: I98256d8af447b4a6b50245c3786e349c8f894b30
Signed-off-by: AudioGod <AudioGod@Sonic-Developers.com>